requirements.txt
fastapi
uvicorn[standard]
httpx
rasterio
numpy
pystac-client
odc-stac
planetary-computer
